added simple_types docs & fixed io funcitons
fixed flush_ not raising an error when the file was closed

+++ libs/nest_co/nest_co.cpp
@@ -295,7 +295,7 @@ NST_FUNC_SIGN(call_)
         NST_SET_CALL_ERROR(nst_sprintf(
             _NST_EM_WRONG_ARG_NUM,
             co->func->arg_num, co->func->arg_num == 1 ? "" : "s",
-            SEQ(co_args)->len, SEQ(co_args)->len == 1 ? "was" : "were"));
+            i64(SEQ(co_args)->len), SEQ(co_args)->len == 1 ? "was" : "were"));
         nst_dec_ref(co_args);
         return nullptr;
     }
